Correct count stats that are credited to the entire defense.
Team stats are given to player ID '0', and we were sometimes missing team defense stats because the first stat under '0' was an offensive stat.

The most interesting team stat we care about is safeties; a safety may be attributed to the entire team if it's entirely the offense's fault, such as a bad snap that goes out of bounds.