Compute scores for the whole week, instead of once per team.
If we trigger the cloud function once for every team, we hit the quota quickly (16 calls per 100 seconds).

Doing it for the whole week seems wasteful because we're computing and writing data for keys that might not have changed. Not sure if that's avoidable or if we should just ignore it.