Differentiate between BQBL season weeks and NFL season weeks.
The list of week IDs in football.js conflates the two. The player standings page uses that list of weeks to show player scores from the BQBL regular season. But the <select> in the nav bar also uses that list of weeks to allow users to view BQBL scores for NFL teams each week, which meant that Week 17 was not selectable.