Fix passer rating points.
Accept either lowercase or uppercase 'passers' key. The 2019 code expects a lowercase key, which is what the NFL JSON scraper produced, but the new scraper produces an uppercase key. Given that a lot of stats have been backfilled, it's easier/safer to make the scoring computation more tolerant of its input than to make the scraper stricter in its output.