Undo the Firebase Promise conversion.
That conversion broke Firebase's live-updating behavior. The `.on` method subscribes to Firebase updates, and calls its callback on every data change. Passing a Promise resolution method as the callback means that the data is only updated once. Future callbacks are discarded because the Promise is already resolved.

This commit effectively reverts the following commits:
 669ce09ee5e571a2ee4474f9304574268fa91e84 (Show an error message to unknown users on the Lineup page.)
 88a4332bd7a03569c491a0a6f00f95a26c1a9a03 (Convert all of our firebase logic to use promises rather than promise-like callbacks)
 6392883d35ca581c0da7e5994eed7756cdae687a (i should actually use language features as intended)
It partially reverts the following commits:
 15e6fa30fa859842ed4279f7153fa461dd0167df: Just the lint changes to firebase.js

(Reverting 669ce09e was necessary because we don't have appropriate error handling on the callback-oriented Firebase methods.)

This commit also applies similar changes to the mock Firebase used in unit tests, which was created after the API change.

It does NOT fix any of the Pro Bowl code, much of which was written after the API change. It's more important to fix current functionality. The Pro Bowl is a problem for somebody to solve at the last minute in 2023.